requirements.txt
streamlit
pyrebase4
pandas
folium
streamlit-folium
